A quick PMF estimation based on alternating Poisson regressions while sharing a dictionary/factors matrix

A quick PMF estimation based on alternating Poisson regressions while sharing a dictionary/factors matrix

Usage

asap_fit_pmf_cbind(
  y_dn_vec,
  maxK,
  max_iter = 100L,
  verbose = TRUE,
  a0 = 1,
  b0 = 1,
  do_log1p = FALSE,
  rseed = 1337L,
  EPS = 1e-08,
  jitter = 1,
  NUM_THREADS = 0L
)

Arguments

y_dn_vec

a list of non-negative data matrices (gene x sample)

maxK

maximum number of factors

max_iter

max number of optimization steps

verbose

verbosity

a0

gamma(a0, b0) default: a0 = 1

b0

gamma(a0, b0) default: b0 = 1

do_log1p

do log(1+y) transformation

rseed

random seed (default: 1337)

EPS

(default: 1e-8)

jitter

(default: 1)

min_iter

min number of optimization steps

Value

a list that contains:

log.likelihood log-likelihood trace

beta dictionary (gene x factor)

log.beta log-dictionary (gene x factor)

log.beta.sd sd(log-dictionary) (gene x factor)

theta a list of loading matrices (sample x factor)

log.theta a list of log loadings (sample x factor)

log.theta.sd a list of standard deviations (sample x factor)